chore: Refactor into a package
