+++ pyproject.toml
@@ -29,6 +29,11 @@ dev-dependencies = [
     "bump-my-version>=0.29.0",
 ]
 
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+
 
 [tool.coverage.run]
 source = ["packages"]
