+++ scripts/fetch_test_data.py
@@ -1,269 +1,13 @@
-import os
-import pathlib
-from abc import ABC, abstractmethod
 from pathlib import Path
-from typing import Any
 
 import pandas as pd
 import pooch
 import xarray as xr
 from intake_esgf import ESGFCatalog
 
+from sample_data import CMIP6Request, DataRequest, Obs4MIPsRequest
+
 OUTPUT_PATH = Path("data")
-
-
-class DataRequest(ABC):
-    """
-    Represents a request for a dataset
-
-    A polymorphic association is used to capture the different types of datasets as each
-    dataset type may have different metadata fields and may need to be handled
-    differently to generate the sample data.
-    """
-
-    def __init__(self, remove_ensembles: bool, time_span: tuple[str, str]):
-        self.remove_ensembles = remove_ensembles
-        self.time_span = time_span
-
-    @abstractmethod
-    def decimate_dataset(self, dataset: xr.Dataset, time_span: tuple[str, str] | None) -> xr.Dataset | None:
-        """Downscale the dataset to a smaller size."""
-        pass
-
-    @abstractmethod
-    def create_out_filename(self, metadata: pd.Series, ds: xr.Dataset, ds_filename: str) -> pathlib.Path:
-        """Create the output filename for the dataset."""
-        pass
-
-
-class CMIP6Request(DataRequest):
-    """
-    Represents a CMIP6 dataset request
-
-    """
-
-    def __init__(self, facets: dict[str, Any], remove_ensembles: bool, time_span: tuple[str, str] | None):
-        self.avail_facets = [
-            "mip_era",
-            "activity_drs",
-            "institution_id",
-            "source_id",
-            "experiment_id",
-            "member_id",
-            "table_id",
-            "variable_id",
-            "grid_label",
-            "version",
-            "data_node",
-        ]
-
-        self.facets = facets
-
-        super().__init__(remove_ensembles, time_span)
-
-        self.cmip6_path_items = [
-            "mip_era",
-            "activity_drs",
-            "institution_id",
-            "source_id",
-            "experiment_id",
-            "member_id",
-            "table_id",
-            "variable_id",
-            "grid_label",
-        ]
-
-        self.cmip6_filename_paths = [
-            "variable_id",
-            "table_id",
-            "source_id",
-            "experiment_id",
-            "member_id",
-            "grid_label",
-        ]
-
-        assert all(key in self.avail_facets for key in self.cmip6_path_items), "Error message"
-        assert all(key in self.avail_facets for key in self.cmip6_filename_paths), "Error message"
-
-    def decimate_dataset(self, dataset: xr.Dataset, time_span: tuple[str, str] | None) -> xr.Dataset | None:
-        """
-        Downscale the dataset to a smaller size.
-
-        Parameters
-        ----------
-        dataset
-            The dataset to downscale
-        time_span
-            The time span to extract from a dataset
-
-        Returns
-        -------
-        xr.Dataset
-            The downscaled dataset
-        """
-        has_latlon = "lat" in dataset.dims and "lon" in dataset.dims
-        has_ij = "i" in dataset.dims and "j" in dataset.dims
-
-        if has_latlon:
-            assert len(dataset.lat.dims) == 1 and len(dataset.lon.dims) == 1
-            result = dataset.interp(lat=dataset.lat[:10], lon=dataset.lon[:10])
-        elif has_ij:
-            # 2d lat/lon grid (generally ocean variables)
-            # Choose a starting point around the middle of the grid to maximise chance that it has values
-            # TODO: Be smarter about this?
-            j_midpoint = len(dataset.j) // 2
-            result = dataset.interp(i=dataset.i[:10], j=dataset.j[j_midpoint : j_midpoint + 10])
-        else:
-            raise ValueError("Cannot decimate this grid: too many dimensions")
-
-        if "time" in dataset.dims and time_span is not None:
-            result = result.sel(time=slice(*time_span))
-            if result.time.size == 0:
-                result = None
-
-        return result
-
-    def create_out_filename(self, metadata: pd.Series, ds: xr.Dataset, ds_filename: str) -> pathlib.Path:
-        """
-        Create the output filename for the dataset.
-
-        Parameters
-        ----------
-        ds
-            Loaded dataset
-
-        Returns
-        -------
-            The output filename
-        """
-        output_path = (
-            Path(os.path.join(*[metadata[item] for item in self.cmip6_path_items]))
-            / f"v{metadata['version']}"
-        )
-        filename_prefix = "_".join([metadata[item] for item in self.cmip6_filename_paths])
-
-        if "time" in ds.dims:
-            time_range = (
-                f"{ds.time.min().dt.strftime('%Y%m').item()}-{ds.time.max().dt.strftime('%Y%m').item()}"
-            )
-            filename = f"{filename_prefix}_{time_range}.nc"
-        else:
-            filename = f"{filename_prefix}.nc"
-
-        return output_path / filename
-
-
-class Obs4MIPsRequest(DataRequest):
-    """
-    Represents a Obs4MIPs dataset request
-
-    """
-
-    def __init__(self, facets: dict[str, Any], remove_ensembles: bool, time_span: tuple[str, str] | None):
-        self.avail_facets = [
-            "activity_id",
-            "institution_id",
-            "source_id",
-            "frequency",
-            "variable_id",
-            "grid_label",
-            "version",
-            "data_node",
-        ]
-
-        self.facets = facets
-
-        super().__init__(remove_ensembles, time_span)
-
-        self.obs4mips_path_items = [
-            "activity_id",
-            "institution_id",
-            "source_id",
-            "variable_id",
-            "grid_label",
-        ]
-
-        self.obs4mips_filename_paths = [
-            "variable_id",
-            "source_id",
-            "grid_label",
-        ]
-
-        assert all(key in self.avail_facets for key in self.obs4mips_path_items), "Error message"
-        assert all(key in self.avail_facets for key in self.obs4mips_filename_paths), "Error message"
-
-    def decimate_dataset(self, dataset: xr.Dataset, time_span: tuple[str, str] | None) -> xr.Dataset | None:
-        """
-        Downscale the dataset to a smaller size.
-
-        Parameters
-        ----------
-        dataset
-            The dataset to downscale
-        time_span
-            The time span to extract from a dataset
-
-        Returns
-        -------
-        xr.Dataset
-            The downscaled dataset
-        """
-        has_latlon = "lat" in dataset.dims and "lon" in dataset.dims
-        has_ij = "i" in dataset.dims and "j" in dataset.dims
-
-        if has_latlon:
-            assert len(dataset.lat.dims) == 1 and len(dataset.lon.dims) == 1
-            result = dataset.interp(lat=dataset.lat[:10], lon=dataset.lon[:10])
-        elif has_ij:
-            # 2d lat/lon grid (generally ocean variables)
-            # Choose a starting point around the middle of the grid to maximise chance that it has values
-            # TODO: Be smarter about this?
-            j_midpoint = len(dataset.j) // 2
-            result = dataset.interp(i=dataset.i[:10], j=dataset.j[j_midpoint : j_midpoint + 10])
-        else:
-            raise ValueError("Cannot decimate this grid: too many dimensions")
-
-        if "time" in dataset.dims and time_span is not None:
-            result = result.sel(time=slice(*time_span))
-            if result.time.size == 0:
-                result = None
-
-        return result
-
-    def create_out_filename(self, metadata: pd.Series, ds: xr.Dataset, ds_filename: str) -> pathlib.Path:
-        """
-        Create the output filename for the dataset.
-
-        Parameters
-        ----------
-        ds
-            Loaded dataset
-
-        Returns
-        -------
-            The output filename
-        """
-        output_path = (
-            Path(os.path.join(*[metadata[item] for item in self.obs4mips_path_items]))
-            / f"v{metadata['version']}"
-        )
-        if ds_filename.name.split("_")[0] == ds.variable_id:
-            filename_prefix = "_".join([metadata[item] for item in self.obs4mips_filename_paths])
-        else:
-            filename_prefix = ds_filename.name.split("_")[0] + "_"
-            filename_prefix += "_".join(
-                [metadata[item] for item in self.obs4mips_filename_paths if item != "variable_id"]
-            )
-
-        if "time" in ds.dims:
-            time_range = (
-                f"{ds.time.min().dt.strftime('%Y%m').item()}-{ds.time.max().dt.strftime('%Y%m').item()}"
-            )
-            filename = f"{filename_prefix}_{time_range}.nc"
-        else:
-            filename = f"{filename_prefix}.nc"
-
-        return output_path / filename
 
 
 def fetch_datasets(request: DataRequest) -> pd.DataFrame:
@@ -272,15 +16,12 @@ def fetch_datasets(request: DataRequest) -> pd.DataFrame:
 
     Parameters
     ----------
-    search_facets
-        Facets to search for
-    remove_ensembles
-        Whether to remove ensembles from the dataset
-        (i.e. include only a single ensemble member)
+    request
+        The request object
 
     Returns
     -------
-    List of paths to the fetched datasets
+        Dataframe that contains metadata and paths to the fetched datasets
     """
     cat = ESGFCatalog()
 
@@ -351,7 +92,7 @@ def create_sample_dataset(request: DataRequest):
             ds_decimated.to_netcdf(output_filename)
 
     # Regenerate the registry.txt file
-    pooch.make_registry(OUTPUT_PATH, "registry.txt")
+    pooch.make_registry(str(OUTPUT_PATH), "registry.txt")
 
 
 if __name__ == "__main__":
