+++ src/sample_data/data_request/__init__.py
@@ -1,0 +1,5 @@
+"""
+Data requests
+
+Provides an abstraction over the different possible data queries that intake-esgf can perform
+"""
